+ map loader sanitize index
