+++ src/Renderer/Batcher2D.cs
@@ -1156,4 +1156,55 @@ namespace ClassicUO.Renderer
             }
         }
     }
+
+
+
+    [StructLayout(LayoutKind.Sequential, Pack = 1)]
+    internal struct VertexPositionColorTexture4 : IVertexType
+    {
+        public const int RealStride = 96;
+
+        VertexDeclaration IVertexType.VertexDeclaration => VertexPositionColorTexture.VertexDeclaration;
+
+        public Vector3 Position0;
+        public Color Color0;
+        public Vector2 TextureCoordinate0;
+        public Vector3 Position1;
+        public Color Color1;
+        public Vector2 TextureCoordinate1;
+        public Vector3 Position2;
+        public Color Color2;
+        public Vector2 TextureCoordinate2;
+        public Vector3 Position3;
+        public Color Color3;
+        public Vector2 TextureCoordinate3;
+    }
+
+
+    [StructLayout(LayoutKind.Sequential, Pack = 1)]
+    internal struct PositionNormalTextureColor : IVertexType
+    {
+
+        public Vector3 Position;
+        public Vector3 Normal;
+        public Vector3 TextureCoordinate;
+        public Vector3 Hue;
+
+        VertexDeclaration IVertexType.VertexDeclaration => VertexDeclaration;
+
+        public static readonly VertexDeclaration VertexDeclaration = new VertexDeclaration(new VertexElement(0, VertexElementFormat.Vector3, VertexElementUsage.Position, 0), // position
+                                                                                           new VertexElement(sizeof(float) * 3, VertexElementFormat.Vector3, VertexElementUsage.Normal, 0), // normal
+                                                                                           new VertexElement(sizeof(float) * 6, VertexElementFormat.Vector3, VertexElementUsage.TextureCoordinate, 0), // tex coord
+                                                                                           new VertexElement(sizeof(float) * 9, VertexElementFormat.Vector3, VertexElementUsage.TextureCoordinate, 1) // hue
+                                                                                         );
+
+        public const int SizeInBytes = sizeof(float) * 12 * 4;
+
+#if DEBUG
+        public override string ToString()
+        {
+            return string.Format("VPNTH: <{0}> <{1}>", Position.ToString(), TextureCoordinate.ToString());
+        }
+#endif
+    }
 }
